Move Netgear LTE entity to its own file
