+++ homeassistant/components/netgear_lte/sensor.py
@@ -8,8 +8,8 @@ from homeassistant.exceptions import PlatformNotReady
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import LTEEntity
 from .const import CONF_SENSOR, DOMAIN
+from .entity import LTEEntity
 from .sensor_types import SENSOR_SMS, SENSOR_SMS_TOTAL, SENSOR_UNITS, SENSOR_USAGE
